Implement retrieving the target element of a mouse event
When using `onClick` and similar events, a user may want to retrieve
the target of the event, for example to compare with other elements in
the DOM tree to see if it's directed to the parent or to some child.
This is done with `target` event field. So implement it.